backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
tenacity>=8.2.3
pillow>=10.3.0
supabase>=2.5.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
alibabacloud_objectdet20191230>=1.3.0
alibabacloud_tea_openapi>=0.3.11
